[AirPlane]Novo formato para setar a direção dos servos;[BatteryMonitor]Ativar a leitura da corrente,retornar watts em tamanho uint32_t;[SwitchFlag]Renomear pasta;[Compass]Adicionar nome ao array;[Declination]Remover variavel desnecessaria;[EscCalibration]Aplique um constrain ao pulso do throttle;[Barometer]adicione retorno as funções,remova calculo matematico para debug,aplique o filtro apenas quando estiver armado;[ParamsToGCS]Utilize a detecção automatica de plataforma para otimizar;[IOMCU]troca de byte para bool;[Build]Envie as informações da placa para o GCS;[SPEED]Adicionar chaves e um enum para as diferentes velocicade dos motores;

+++ src/BatteryMonitor/BATTERY.h
@@ -30,7 +30,7 @@ public:
   void Read_Voltage(void);
   void Read_Current(void);
   void Calculate_Total_Mah(void);
-  uint16_t GetWatts();
+  uint32_t GetWatts();
 
 private:
   uint8_t BattMinVoltageSelect;
